Done explicit type conversion to get rid of compiler warnings for some compilers.

+++ code-experiments/test/unit-test/test_coco_archive.c
@@ -41,8 +41,8 @@ static void test_coco_archive(void **state) {
       break;
 
     /* Add solution to the archive */
-    line = coco_strdupf("%lu\t%f\t%f\t%f\t%f\t%f\n", number_of_evaluations, x[0], x[1], y[0], y[1],
-        hypervolume_read);
+    line = coco_strdupf("%lu\t%f\t%f\t%f\t%f\t%f\n", (unsigned long) number_of_evaluations, x[0], x[1], y[0],
+    		y[1], hypervolume_read);
     coco_archive_add_solution(archive, y[0], y[1], line);
     coco_free_memory(line);
   }
